Add delete confirmation

+++ public/script.js
@@ -12,7 +12,9 @@ fetchData().then((reviews) => {
     const btnDel = document.createElement("button");
     btnDel.innerHTML = `x`;
     btnDel.addEventListener("click", async () => {
-      const response = await fetch(`http://localhost:3333/reviews/`, {
+      const deleteConfirmation = confirm("まじで? (｡╯︵╰｡)");
+      if (!deleteConfirmation) return;
+      await fetch(`http://localhost:3333/reviews/`, {
         method: "delete",
         mode: "cors",
         headers: {
@@ -22,7 +24,7 @@ fetchData().then((reviews) => {
           kanji: item.kanji,
         }),
       }).then((res) => console.log(res));
-      location.reload(); 
+      location.reload();
     });
     btnDel.classList.add("btnDel");
     li.innerHTML = `<a href="https://jisho.org/search/${item.kanji}" target="_blank" rel="noopener">${item.kanji}</a> 「${item.kana}」`;
